refactor(trpc): split monolithic router into modular procedures
