+++ src/trpc/routers/_app.ts
@@ -1,7 +1,13 @@
+import { coursesRoute } from "@/app/modules/courses/server/procedures/course";
 import { createTRPCRouter } from "../init";
-import { coursesRoute } from "@/app/modules/courses/server/procedures";
+import { chaptersRoute } from "@/app/modules/courses/server/procedures/chapters";
+import { attachmentsRoute } from "@/app/modules/courses/server/procedures/attachments";
+import { categoriesRoute } from "@/app/modules/courses/server/procedures/category";
 export const appRouter = createTRPCRouter({
   courses: coursesRoute,
+  chapters: chaptersRoute,
+  attachments: attachmentsRoute,
+  categories: categoriesRoute,
 });
 // export type definition of API
 export type AppRouter = typeof appRouter;
